Added Time mode .

+++ Quartett/app/src/main/java/de/in/uulm/map/quartett/data/LocalGameState.java
@@ -19,7 +19,7 @@ public class LocalGameState extends SugarRecord {
     public int mUserPoints;
     public int mAIPoints;
     public int mCurrentRound;
-    public int mCurrentTimeInMillis;
+    public long mCurrentTimeInMillis;
     public long mGameTimeInMillis;
     public int mMaxPoints;
     public int mMaxRounds;
